refs #124 Move mastodon api methods under mastodon dir
