+++ src/mastodon.ts
@@ -1,4 +1,4 @@
-import APIClient from './api_client'
+import APIClient from './mastodon/api_client'
 import { ProxyConfig } from './proxy_config'
 import OAuth from './oauth'
 import Response from './response'
@@ -26,8 +26,8 @@ import { Activity } from './entities/activity'
 import { Emoji } from './entities/emoji'
 import { PushSubscription } from './entities/push_subscription'
 import { Token } from './entities/token'
-import StreamListener from './stream_listener'
-import WebSocket from './web_socket'
+import StreamListener from './mastodon/stream_listener'
+import WebSocket from './mastodon/web_socket'
 
 const NO_REDIRECT = 'urn:ietf:wg:oauth:2.0:oob'
 const DEFAULT_URL = 'https://mastodon.social'
