+++ src/mastodon/api_client.ts
@@ -3,10 +3,10 @@ import { OAuth2 } from 'oauth'
 
 import StreamListener from './stream_listener'
 import WebSocket from './web_socket'
-import Response from './response'
-import { RequestCanceledError } from './cancel'
-import proxyAgent, { ProxyConfig } from './proxy_config'
-import OAuth from './oauth'
+import Response from '../response'
+import { RequestCanceledError } from '../cancel'
+import proxyAgent, { ProxyConfig } from '../proxy_config'
+import OAuth from '../oauth'
 
 const NO_REDIRECT = 'urn:ietf:wg:oauth:2.0:oob'
 const DEFAULT_URL = 'https://mastodon.social'
