+++ src/mastodon/parser.ts
@@ -1,7 +1,7 @@
 import { EventEmitter } from 'events'
-import { Status } from './entities/status'
-import { Notification } from './entities/notification'
-import { Conversation } from './entities/conversation'
+import { Status } from '../entities/status'
+import { Notification } from '../entities/notification'
+import { Conversation } from '../entities/conversation'
 
 /**
  * Parser
