+++ src/mastodon/stream_listener.ts
@@ -2,10 +2,10 @@ import { EventEmitter } from 'events'
 import axios, { CancelTokenSource, AxiosRequestConfig } from 'axios'
 import httpAdapter from 'axios/lib/adapters/http'
 import Parser from './parser'
-import { Status } from './entities/status'
-import { Notification } from './entities/notification'
-import { Conversation } from './entities/conversation'
-import proxyAgent, { ProxyConfig } from './proxy_config'
+import { Status } from '../entities/status'
+import { Notification } from '../entities/notification'
+import { Conversation } from '../entities/conversation'
+import proxyAgent, { ProxyConfig } from '../proxy_config'
 
 const STATUS_CODES_TO_ABORT_ON: Array<number> = [400, 401, 403, 404, 406, 410, 422]
