+++ test/integration/cancelWorker.ts
@@ -1,4 +1,4 @@
-import APIClient from '@/api_client'
+import APIClient from '@/mastodon/api_client'
 
 export function cancel(client: APIClient) {
   return client.cancel()
